+++ ReleaseNotes.md
@@ -10,6 +10,8 @@
 - Item sheets are flavored to the kind of Actor owner (assets by @Sylph)
 - Dice so nice integration: roll botch dice displayed.
 - Basic aging roll (no effect yet)
+- Fatigue management rework
+- New active effect to add fatigue levels
 
 ### Bug fixes